requirements.txt
apache-airflow>=2.7.0
pydantic>=2.0
mysql-connector-python>=8.2.0
httpx>=0.27.0
databricks-sql-connector>=3.0.0
dbt-databricks>=1.7.0
